fix(death): phase 5 — extract utilities to DeathUtils.ts (~1605 lines)
Extract pure functions from PlayerDeathSystem to DeathUtils.ts:
- sanitizeKilledBy, getItemValue, splitItemsForSafeDeath
- validatePosition, isPositionInBounds, isValidPositionNumber
- ITEMS_KEPT_ON_DEATH, POSITION_VALIDATION constants

Fix splitItemsForSafeDeath stack explosion: no longer expands stacked
items (qty=10000 → 10000 array entries). Now operates on unique item
slots with greedy quantity assignment — O(n log n) on unique items.

PlayerDeathSystem: 2083 → 1605 lines
